Switch to using setup.cfg
Running `pip install .` or `python -m setup.py install` runs into import errors if the package dependencies are not met. This project lacks a requirements.txt file, and the README doesn't specify the dependencies either, so it's an extra undocumented step to complete before installing.

Because the package __init__.py imports all from its submodules, importing the package requires the dependencies to be installed, which causes an ImportError when running setup.py, because it imports the package.

Abstracting all of the setuptools options into a setup.cfg file without referencing the package attributes resolves the issue, and allows the package to import all from the submodules and maintain package metadata attributes for referencing via other code.

Additionally, a setup.cfg file is generally easier to maintain than direct setup() keyword arguments.